fix migration alter date to integer (#500)

+++ django_project/population_data/models.py
@@ -47,7 +47,7 @@ class NatureOfPopulation(models.Model):
 
 class PopulationCountAbstract(models.Model):
     """"Population count abstract model."""
-    year = models.PositiveBigIntegerField()
+    year = models.PositiveIntegerField()
     owned_species = models.ForeignKey('species.OwnedSpecies', on_delete=models.CASCADE)
     total = models.IntegerField()
     adult_male = models.IntegerField(null=True, blank=True)
